# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/AddSinglePersonAndMatch": {
            "post": {
                "description": "Add a new user to the matching system and find any possible matches for the new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "註冊帳號",
                "parameters": [
                    {
                        "description": "要註冊的帳號",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.C2S_Register"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.S2C_Login"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    }
                }
            }
        },
        "/v1/QuerySinglePeople": {
            "post": {
                "description": "QuerySinglePeople : Find the most N possible matched single people",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "尋找最多 N 個可能匹配的單身人士",
                "parameters": [
                    {
                        "description": "要匹配的單身人士",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserQueryCheck"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.S2C_MatchPeople"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    }
                }
            }
        },
        "/v1/RemoveSinglePerson": {
            "delete": {
                "description": "Remove a user from the matching system so that the user cannot be matched anymore",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "移除帳號",
                "parameters": [
                    {
                        "description": "要檢查的帳號",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserCheck"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.S2C_Login"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.C2S_Register": {
            "type": "object",
            "properties": {
                "gender": {
                    "description": "性別",
                    "type": "integer"
                },
                "height": {
                    "description": "身高",
                    "type": "integer"
                },
                "password": {
                    "description": "用戶密碼",
                    "type": "string"
                },
                "username": {
                    "description": "姓名",
                    "type": "string"
                }
            }
        },
        "model.C2S_Register_Base": {
            "type": "object",
            "properties": {
                "gender": {
                    "description": "性別",
                    "type": "integer"
                },
                "height": {
                    "description": "身高",
                    "type": "integer"
                },
                "username": {
                    "description": "姓名",
                    "type": "string"
                }
            }
        },
        "model.S2C_Login": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.S2C_MatchPeople": {
            "type": "object",
            "properties": {
                "teamMemberA": {
                    "$ref": "#/definitions/model.C2S_Register_Base"
                },
                "teamMemberB": {
                    "$ref": "#/definitions/model.C2S_Register_Base"
                },
                "team_index": {
                    "description": "MatchPeopleList []MatchPeople `json:\"match_people_list\"`",
                    "type": "integer"
                }
            }
        },
        "model.UserCheck": {
            "type": "object",
            "properties": {
                "userId": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.UserQueryCheck": {
            "type": "object",
            "properties": {
                "gender": {
                    "description": "性別",
                    "type": "integer"
                },
                "height": {
                    "description": "身高",
                    "type": "integer"
                },
                "needCount": {
                    "description": "需要人數",
                    "type": "integer"
                },
                "username": {
                    "description": "用戶名",
                    "type": "string"
                }
            }
        },
        "response.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        }
    }
}